Move the logic for updating bot stats to a util
