+++ src/Utils/index.ts
@@ -1,18 +1,18 @@
-export * from './BuildPagination'
-export * from './CanvasHelper'
-export * from './CommandCategories'
-export * from './EmbedError'
-export * from './Footer'
-export * from './GraphQLRequest'
-export * from './Logging'
-export * from './RSACryption'
-export * from './getOptions'
-export * from './types'
-export * from './registerCommands'
-export * from './registerEvents'
-export * from './registerComponents'
-export * from './getAnilistUser'
-export * from './getSubcommand';
-export * from './handleMediaData';
-export * from './common'
-export * from './getStats'
+export * from "./BuildPagination";
+export * from "./CanvasHelper";
+export * from "./CommandCategories";
+export * from "./EmbedError";
+export * from "./Footer";
+export * from "./GraphQLRequest";
+export * from "./Logging";
+export * from "./RSACryption";
+export * from "./getOptions";
+export * from "./types";
+export * from "./registerCommands";
+export * from "./registerEvents";
+export * from "./registerComponents";
+export * from "./getAnilistUser";
+export * from "./getSubcommand";
+export * from "./handleMediaData";
+export * from "./common";
+export * from "./botStats";
